refactor(utils): Move metrics/get-metric-changed to report
