+++ packages/utils/src/metrics/index.js
@@ -1,2 +1,1 @@
-export * from './get-metric-changed';
 export * from './get-metric-type';
