+++ packages/utils/src/report/index.js
@@ -1,4 +1,5 @@
-export * from './create';
-export * from './compare-metrics';
 export * from './add-row-data';
+export * from './compare-metrics';
+export * from './create';
+export * from './get-metric-changed';
 export * from './merge-metrics-by-key';
